pyproject.toml
[tool.poetry]
name = "maple-cody-recommandation-system"
version = "0.1.0"
description = ""
authors = ["leecs0503"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyppeteer = "^1.0.2"
pandas = "^1.4.4"
bs4 = "^0.0.1"
requests = "^2.28.1"
PySocks = "^1.7.1"
aiohttp = "^3.8.3"
Pillow = "^9.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
flake8 = "^5.0.4"
isort = "^5.10.1"
black = "^22.8.0"
pyright = "^1.1.270"
ipykernel = "^6.15.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = ".venv"
